requirements.txt
flask
flask-sqlalchemy
werkzeug
cryptography
argon2-cffi
pytest